Move gradient widgets to ui
