+++ src/ui/gradientwidgets/gradientslistwidget.cpp
