+++ src/ui/widgets/colorwidgetshaders.cpp
