+++ src/ui/widgets/colorwidgetshaders.h
